+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `lang-util`: Add lang_util::error::ResolvedPosition::without_source_number
 
 ## [0.1.3] - 2021-04-16
 ### Added